requirements.txt
annotated-types==0.6.0
anyio==3.7.0
bcrypt==4.1.2
click==8.1.3
colorama==0.4.6
dnspython==2.6.0
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httptools==0.5.0
idna==3.4
pydantic==1.10.8
pydantic_core==2.16.2
pymongo==4.6.1
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
